Update GMP target to include a directory property.
Once libgmp is found, make sure to record the directory in which it was found so that
LDFLAGS can be updated (later) when calling `configure` for the library. This makes
it possible to build the project on platforms where `libgmp.a` may be in a place such
as `/usr/local/lib`, perhaps.